Don't cast float to int in UnitChartValueFormatter and fix temperature formatting in WeatherBlockSubcomponent

+++ weather/app/src/main/java/com/pelmenstar/projktSens/weather/app/formatters/UnitChartValueFormatter.java
@@ -19,6 +19,6 @@ public final class UnitChartValueFormatter extends ValueFormatter {
     @NotNull
     @Override
     public String formatToString(float value) {
-        return unitFormatter.formatValue((int) value, unit);
+        return unitFormatter.formatValue(value, unit);
     }
 }
